Updates FastAPI version and steps for creating the workflow.
Thanks to [imtiaz-qureshi](https://github.com/imtiaz-qureshi) for
sharing that the FastAPI version pinned in the requirements file was
causing the workflow to fail.  I updated the version along with uvicorn
as well.

I also updated the README file to be in line with the current steps
needed to create a workflow in the GitHub Actions Web UI.